requirements.txt
Django==3.2.12
SQLAlchemy==1.4.34
Pillow==9.1.0
mysql-connector==2.2.9
APScheduler==3.9.1
requests==2.27.1
